fix(backend): resolve daily-summary endpoint 500 errors
- Fix field name mismatch between service and response model
- Update daily_summary_service to return correct field names (daily_goal, total_consumed, remaining)
- Add LoggedMealEntry model to match service output structure
- Convert float values to int for compatibility with existing model
- Remove unused fields (percentage, goal_exceeded, excess_calories, entry_date)
- Fix int() casting issues in API endpoints for calories fields
- Resolves 500 Internal Server Error on /api/v1/users/me/daily-summary endpoint
- Dashboard now loads correctly with proper data structure

+++ backend/app/api/users.py
@@ -1089,7 +1089,7 @@ async def set_nutrition_goal(
 
     # TEMPORARY: Hardcoded user_id for development
     user_id = 1
-    
+
     # Find or create calorie goal
     calorie_goal = db.query(DailyCalorieGoal).filter(
         DailyCalorieGoal.user_id == user_id
@@ -1146,7 +1146,7 @@ async def set_nutrition_goal(
         LoggedMealSummary(
             log_id=m.id,
             name=m.name,
-            calories=int(m.calories or 0),
+            calories=float(m.calories or 0),
             meal_type=m.meal_type or "Unknown",
             logged_at=m.updated_at.isoformat() if m.updated_at else datetime.now(UTC).isoformat()
         )
